requirements.txt
# Flask==3.0.3
gunicorn==23.0.0

Flask==2.0.3
Flask-Admin==1.6.1
Flask-AdminLTE3==1.0.9
Flask-WTF==1.2.1
email-validator==2.2.0
Werkzeug==2.0.3
flask_babelex==0.9.4
pytz==2023.3

Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.9.10
SQLAlchemy==1.4.46
Flask-Mail==0.10.0
itsdangerous==2.2.0
bcrypt==4.2.0
Flask-Login==0.6.1
pyotp==2.9.0
tink==1.10.0

minio==7.1.14

wtforms_sqlalchemy==0.4.2

fpdf==1.7.2
Loguru==0.7.3
google-genai